Move try_populate_devicetree to pyanaconda.ui.lib
Remove the devicetree argument from the function try_populate_devicetree and
move it to pyanaconda.ui.lib. Don't call this function from the Storage module.
