+++ pyanaconda/ui/gui/spokes/advstorage/dasd.py
@@ -21,7 +21,7 @@ from pyanaconda.modules.common.constants.services import STORAGE
 from pyanaconda.modules.common.errors.configuration import StorageDiscoveryError
 from pyanaconda.modules.common.task import async_run_task
 from pyanaconda.ui.gui import GUIObject
-from pyanaconda.storage.utils import try_populate_devicetree
+from pyanaconda.ui.lib.storage import try_populate_devicetree
 
 __all__ = ["DASDDialog"]
 
@@ -65,7 +65,7 @@ class DASDDialog(GUIObject):
         # We need to call this to get the device nodes to show up
         # in our devicetree.
         if self._update_devicetree:
-            try_populate_devicetree(self._storage.devicetree)
+            try_populate_devicetree()
         return rc
 
     def on_start_clicked(self, *args):
