+++ pyanaconda/ui/lib/storage.py
@@ -187,3 +187,26 @@ def unmark_protected_device(spec):
         protected_devices.remove(spec)
 
     disk_selection_proxy.SetProtectedDevices(protected_devices)
+
+
+def try_populate_devicetree():
+    """Try to populate a device tree.
+
+    Try to populate the devic etree while catching errors and dealing with
+    some special ones in a nice way (giving user chance to do something about
+    them).
+    """
+    device_tree = STORAGE.get_proxy(DEVICE_TREE)
+
+    while True:
+        try:
+            task_path = device_tree.FindDevicesWithTask()
+            task_proxy = STORAGE.get_proxy(task_path)
+            sync_run_task(task_proxy)
+        except StorageError as e:
+            if error_handler.cb(e) == ERROR_RAISE:
+                raise
+            else:
+                continue
+        else:
+            break
